requirements.txt
# Core dependencies
streamlit==1.32.0
python-dotenv==1.0.1

# OCR and document processing
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
PyPDF2==3.0.1
pdfplumber==0.11.0

# LLM integration
ollama==0.1.7
langchain==0.1.11
langchain-community==0.0.28

# Data validation and schemas
pydantic==2.6.3
pydantic-settings==2.2.1

# Security and encryption
cryptography==42.0.5
sqlcipher3==0.5.2

# Utilities
python-dateutil==2.9.0
pytz==2024.1
typing-extensions==4.10.0

# Testing
pytest==8.1.1
pytest-cov==5.0.0
pytest-mock==3.12.0

# Code quality
black==24.3.0
flake8==7.0.0
mypy==1.9.0
